+++ sdk/src/main/java/com/herewhite/sdk/WhiteSdk.java
@@ -11,6 +11,10 @@ import com.herewhite.sdk.domain.Promise;
 import com.herewhite.sdk.domain.RoomPhase;
 import com.herewhite.sdk.domain.SDKError;
 import com.herewhite.sdk.domain.UrlInterrupter;
+import com.herewhite.sdk.internal.PlayerJsInterfaceImpl;
+import com.herewhite.sdk.internal.RoomJsInterfaceImpl;
+import com.herewhite.sdk.internal.RtcJsInterfaceImpl;
+import com.herewhite.sdk.internal.SdkJsInterfaceImpl;
 
 import org.json.JSONObject;
 
@@ -36,10 +40,10 @@ public class WhiteSdk {
      * <p>
      * SDK 通过 `CommonCallbacks` 类向 app 报告 SDK 运行时的各项事件。
      *
-     * @param commonCallbacks 通用事件回调，详见 {@link CommonCallbacks CommonCallbacks}。
+     * @param commonCallback 通用回调事件，详见 {@link CommonCallback CommonCallback}
      */
-    public void setCommonCallbacks(CommonCallbacks commonCallbacks) {
-        sdkJsInterface.setCommonCallbacks(commonCallbacks);
+    public void setCommonCallbacks(CommonCallback commonCallback) {
+        sdkJsInterface.setCommonCallbacks(commonCallback);
     }
 
     private final boolean onlyCallbackRemoteStateModify;
@@ -62,7 +66,7 @@ public class WhiteSdk {
      * @return 当前的 SDK 版本号，格式为字符串，如 `"2.12.11"`。
      */
     public static String Version() {
-        return "2.12.11";
+        return "2.12.26";
     }
 
     /**