+++ sdk/src/main/java/com/herewhite/sdk/WhiteSdkConfiguration.java
@@ -104,6 +104,10 @@ public class WhiteSdkConfiguration extends WhiteObject {
             this.scheme = scheme;
         }
         /// @endcond
+  
+        public  PptParams() {
+
+        }
     }