+++ sdk/src/main/java/com/herewhite/sdk/converter/ConvertType.java
@@ -1,0 +1,6 @@
+package com.herewhite.sdk.converter;
+
+public enum ConvertType {
+    Static,
+    Dynamic
+}