+++ sdk/src/test/java/com/herewhite/sdk/domain/RegionTest.java
@@ -22,6 +22,9 @@ public class RegionTest {
 
     private String testCN = "{\"region\":\"cn-hz\"}";
     private String testUS = "{\"region\":\"us-sv\"}";
+    private String testSG = "{\"region\":\"sg\"}";
+    private String testIN = "{\"region\":\"in-mum\"}";
+    private String testGB = "{\"region\":\"gb-lon\"}";
 
     @Test
     public void serialize() {
@@ -33,8 +36,14 @@ public class RegionTest {
         foobar.setRegion(Region.us);
         assertEquals(testUS, new Gson().toJson(foobar));
 
+        foobar.setRegion(Region.sg);
+        assertEquals(testSG, new Gson().toJson(foobar));
 
-        new Gson().toJson(Region.cn);
+        foobar.setRegion(Region.in_mum);
+        assertEquals(testIN, new Gson().toJson(foobar));
+
+        foobar.setRegion(Region.gb_lon);
+        assertEquals(testGB, new Gson().toJson(foobar));
     }
 
     @Test
@@ -42,5 +51,11 @@ public class RegionTest {
         assertEquals(Region.cn, new Gson().fromJson(testCN, Foobar.class).region);
 
         assertEquals(Region.us, new Gson().fromJson(testUS, Foobar.class).region);
+
+        assertEquals(Region.sg, new Gson().fromJson(testSG, Foobar.class).region);
+
+        assertEquals(Region.in_mum, new Gson().fromJson(testIN, Foobar.class).region);
+
+        assertEquals(Region.gb_lon, new Gson().fromJson(testGB, Foobar.class).region);
     }
 }